refactor: accessing services through service container in symfony bundles
