--- src/bridge/symfony/postgresql-bundle/src/Flow/Bridge/Symfony/PostgreSqlBundle/DependencyInjection/Configuration.php
@@ -1,197 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Flow\Bridge\Symfony\PostgreSqlBundle\DependencyInjection;
-
-use Symfony\Component\Config\Definition\Builder\TreeBuilder;
-use Symfony\Component\Config\Definition\ConfigurationInterface;
-
-final class Configuration implements ConfigurationInterface
-{
-    public function getConfigTreeBuilder() : TreeBuilder
-    {
-        $treeBuilder = new TreeBuilder('flow_postgresql');
-
-        $treeBuilder->getRootNode()
-            ->children()
-                ->arrayNode('connections')
-                    ->requiresAtLeastOneElement()
-                    ->useAttributeAsKey('name')
-                    ->arrayPrototype()
-                        ->children()
-                            ->scalarNode('dsn')
-                                ->isRequired()
-                                ->cannotBeEmpty()
-                                ->info('PostgreSQL connection DSN (e.g. postgresql://user:pass@localhost:5432/dbname)')
-                            ->end()
-                            ->booleanNode('test_transaction_rollback')
-                                ->defaultFalse()
-                                ->info('When true and flow-php/phpunit-postgresql-bridge is installed, wraps the connection with StaticClient for transaction rollback in tests.')
-                            ->end()
-                            ->arrayNode('context')
-                                ->info('Extra key/value pairs merged into the Flow\\PostgreSql\\Client\\Context for every mapper call. Values can be literals, @service_id references, %parameter% placeholders, or %env(VAR)% expressions.')
-                                ->useAttributeAsKey('name')
-                                ->variablePrototype()->end()
-                            ->end()
-                            ->arrayNode('telemetry')
-                                ->children()
-                                    ->scalarNode('service_id')
-                                        ->isRequired()
-                                        ->cannotBeEmpty()
-                                        ->info('Service ID of the Telemetry instance (e.g. flow.telemetry)')
-                                    ->end()
-                                    ->scalarNode('clock_service_id')
-                                        ->defaultNull()
-                                        ->info('Service ID of a PSR ClockInterface implementation. Default: creates SystemClock')
-                                    ->end()
-                                    ->booleanNode('trace_queries')->defaultTrue()->end()
-                                    ->booleanNode('trace_transactions')->defaultTrue()->end()
-                                    ->booleanNode('collect_metrics')->defaultTrue()->end()
-                                    ->booleanNode('log_queries')->defaultFalse()->end()
-                                    ->integerNode('max_query_length')->defaultValue(1000)->min(0)->end()
-                                    ->booleanNode('include_parameters')->defaultFalse()->end()
-                                    ->integerNode('max_parameters')->defaultValue(10)->min(0)->end()
-                                    ->integerNode('max_parameter_length')->defaultValue(100)->min(0)->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                    ->end()
-                ->end()
-                ->arrayNode('messenger')
-                    ->info('Enables the Symfony Messenger PostgreSQL transport. Requires flow-php/symfony-postgresql-messenger-bridge.')
-                    ->canBeEnabled()
-                    ->children()
-                        ->scalarNode('table_name')
-                            ->defaultValue('messenger_messages')
-                            ->cannotBeEmpty()
-                            ->info('Name of the table that stores messenger messages.')
-                        ->end()
-                        ->scalarNode('schema')
-                            ->defaultValue('public')
-                            ->cannotBeEmpty()
-                            ->info('Schema that owns the messenger table.')
-                        ->end()
-                    ->end()
-                ->end()
-                ->arrayNode('cache')
-                    ->info('Defines PostgreSQL-backed Symfony Cache pools. Requires flow-php/symfony-postgresql-cache-bridge.')
-                    ->addDefaultsIfNotSet()
-                    ->children()
-                        ->arrayNode('pools')
-                            ->info('Named cache pools. Each becomes a service "flow.postgresql.cache.pool.<name>" usable as adapter: <id> in framework.cache.pools.')
-                            ->useAttributeAsKey('name')
-                            ->arrayPrototype()
-                                ->children()
-                                    ->scalarNode('connection')
-                                        ->defaultNull()
-                                        ->info('flow_postgresql.connections key to use. Defaults to the first declared connection when null.')
-                                    ->end()
-                                    ->scalarNode('table_name')->defaultValue('cache_items')->cannotBeEmpty()->end()
-                                    ->scalarNode('schema')->defaultValue('public')->cannotBeEmpty()->end()
-                                    ->scalarNode('id_col')->defaultValue('item_id')->cannotBeEmpty()->end()
-                                    ->scalarNode('data_col')->defaultValue('item_data')->cannotBeEmpty()->end()
-                                    ->scalarNode('lifetime_col')->defaultValue('item_lifetime')->cannotBeEmpty()->end()
-                                    ->scalarNode('time_col')->defaultValue('item_time')->cannotBeEmpty()->end()
-                                    ->scalarNode('namespace')
-                                        ->defaultValue('')
-                                        ->info('Cache pool namespace. Allowed chars: -+.A-Za-z0-9')
-                                    ->end()
-                                    ->integerNode('default_lifetime')->defaultValue(0)->min(0)->end()
-                                    ->scalarNode('marshaller_service_id')->defaultNull()->end()
-                                    ->booleanNode('share_connection')
-                                        ->defaultFalse()
-                                        ->info('When true, the pool reuses the named connection\'s Client instead of opening its own pg_connect. Off by default.')
-                                    ->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                    ->end()
-                ->end()
-                ->arrayNode('session')
-                    ->info('Defines the PostgreSQL-backed Symfony Session handler. Requires flow-php/symfony-postgresql-session-bridge.')
-                    ->canBeEnabled()
-                    ->children()
-                        ->scalarNode('connection')
-                            ->defaultNull()
-                            ->info('flow_postgresql.connections key to use. Defaults to the first declared connection when null.')
-                        ->end()
-                        ->scalarNode('table_name')->defaultValue('sessions')->cannotBeEmpty()->end()
-                        ->scalarNode('schema')->defaultValue('public')->cannotBeEmpty()->end()
-                        ->scalarNode('id_col')->defaultValue('sess_id')->cannotBeEmpty()->end()
-                        ->scalarNode('data_col')->defaultValue('sess_data')->cannotBeEmpty()->end()
-                        ->scalarNode('lifetime_col')->defaultValue('sess_lifetime')->cannotBeEmpty()->end()
-                        ->scalarNode('time_col')->defaultValue('sess_time')->cannotBeEmpty()->end()
-                        ->enumNode('lock_mode')
-                            ->values(['none', 'advisory', 'transactional'])
-                            ->defaultValue('transactional')
-                            ->info('Locking strategy. "transactional" uses SELECT FOR UPDATE; "advisory" uses pg_advisory_lock; "none" disables locking.')
-                        ->end()
-                        ->integerNode('ttl')
-                            ->defaultNull()
-                            ->min(0)
-                            ->info('Session lifetime in seconds. When null, falls back to ini "session.gc_maxlifetime".')
-                        ->end()
-                        ->booleanNode('share_connection')
-                            ->defaultFalse()
-                            ->info('When true, the handler reuses the named connection\'s Client instead of opening its own pg_connect. Off by default.')
-                        ->end()
-                    ->end()
-                ->end()
-                ->arrayNode('migrations')
-                    ->canBeEnabled()
-                    ->children()
-                        ->scalarNode('directory')
-                            ->defaultValue('%kernel.project_dir%/migrations')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('namespace')
-                            ->defaultValue('App\\Migrations')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('table_name')
-                            ->defaultValue('flow_migrations')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('table_schema')
-                            ->defaultValue('public')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('migration_file_name')
-                            ->defaultValue('migration.php')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->scalarNode('rollback_file_name')
-                            ->defaultValue('rollback.php')
-                            ->cannotBeEmpty()
-                        ->end()
-                        ->booleanNode('all_or_nothing')
-                            ->defaultFalse()
-                            ->info('Wrap all migrations in a single transaction (default: false)')
-                        ->end()
-                        ->booleanNode('generate_rollback')
-                            ->defaultTrue()
-                            ->info('Generate rollback files when creating migrations (default: true)')
-                        ->end()
-                    ->end()
-                ->end()
-                ->arrayNode('catalog_providers')
-                    ->info('List of catalog providers to merge into the target schema. Each entry must have either "catalog_provider_id" or "catalog".')
-                    ->arrayPrototype()
-                        ->children()
-                            ->scalarNode('catalog_provider_id')
-                                ->defaultNull()
-                                ->info('Service ID of Flow\\PostgreSql\\Schema\\CatalogProvider')
-                            ->end()
-                            ->variableNode('catalog')
-                                ->defaultNull()
-                                ->info('Inline catalog definition matching Catalog::fromArray() shape')
-                            ->end()
-                        ->end()
-                    ->end()
-                ->end()
-            ->end();
-
-        return $treeBuilder;
-    }
-}
